install process improvements

+++ install/emonpilcd.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+source config.ini
 
 sudo apt-get install -y python-smbus i2c-tools python-rpi.gpio python-gpiozero
 sudo pip install xmltodict
@@ -10,19 +11,22 @@ sudo sed -i -n '/i2c-dev/!p;$a i2c-dev' /etc/modules
 
 # emonPiLCD folder is symlinked to /usr/share
 # service expects python script to be at this location
-sudo ln -s $usrdir/emonpi/lcd/ /usr/share/emonPiLCD
+if [ ! -L /usr/share/emonPiLCD ]; then
+    sudo ln -s $usrdir/emonpi/lcd/ /usr/share/emonPiLCD
+fi
 
-# emonPiLCD Logger
-sudo mkdir /var/log/emonpilcd
-# Permissions?
-touch /var/log/emonpilcd/emonpilcd.log
+
+if [ ! -d /var/log/emonpilcd ]; then
+    # emonPiLCD Logger
+    sudo mkdir /var/log/emonpilcd
+    # Permissions?
+    sudo touch /var/log/emonpilcd/emonpilcd.log
+fi
 
 # EmonBase / EmonPi flags are stored in /home/pi/data
 cd
-mkdir data
+if [ ! -d data ]; then mkdir data; fi
 
 # Install and start emonPiLCD service
-sudo ln -s /usr/share/emonPiLCD/emonPiLCD.service /lib/systemd/system
-sudo systemctl daemon-reload
-sudo systemctl enable emonPiLCD.service
-sudo systemctl start emonPiLCD.service
+servicepath=/usr/share/emonPiLCD/emonPiLCD.service
+$usrdir/emonpi/update/install_emoncms_service.sh $servicepath emonPiLCD
